server/requirements.txt
albumentations==1.4.4
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
blinker==1.8.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
einops==0.8.0
entmax==1.3
exceptiongroup==1.2.1
fastapi==0.110.3
filelock==3.14.0
Flask==3.0.3
Flask-Cors==4.0.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.22.2
idna==3.7
imageio==2.34.1
imagesize==1.4.1
importlib_metadata==7.1.0
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lazy_loader==0.4
Levenshtein==0.25.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
munch==4.0.0
networkx==3.2.1
numpy==1.26.4
opencv-python-headless==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pix2tex==0.1.2
protobuf==4.25.3
pyarrow==16.0.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.0
Pygments==2.17.2
pylatexenc==2.10
pynput==1.7.6
PyQt6==6.7.0
PyQt6-Qt6==6.7.0
PyQt6-sip==13.6.0
PyQt6-WebEngine==6.7.0
PyQt6-WebEngine-Qt6==6.7.0
pyreadline3==3.4.1
PySide6==6.7.0
PySide6_Addons==6.7.0
PySide6_Essentials==6.7.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.25.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.8.1
referencing==0.35.0
regex==2024.4.28
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.3
scikit-image==0.22.0
scikit-learn==1.4.2
scipy==1.13.0
screeninfo==0.8.1
shiboken6==6.7.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.37.2
streamlit==1.33.0
sympy==1.12
tbb==2021.12.0
tenacity==8.2.3
threadpoolctl==3.5.0
tifffile==2024.4.24
timm==0.5.4
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.0
torchtext==0.18.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.2
transformers==4.40.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
watchdog==4.0.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.2
x-transformers==0.15.0
zipp==3.18.1
